refactor(chatbot): extract formatRelativeTime to shared utility
